+++ docs/Changelog.md
@@ -18,6 +18,7 @@ Better handling of multiple schedules in batch periodic engine - fixed running o
 * [#1422](https://github.com/TouK/nussknacker/pull/1422) Remove `ServiceReturningType` and `WithExplicitMethod`, added helpers, small refactor
 * [#2278](https://github.com/TouK/nussknacker/pull/1422) SQL Variable is removed          
 * [#2280](https://github.com/TouK/nussknacker/pull/2280) Default values for parameters can be setup programmatically now - thanks to `@DefaultValue` annotation and `Parameter.defaultValue` field.
+* [#2293](https://github.com/TouK/nussknacker/pull/2293) Enhancement: change `nodeCategoryMapping` configuration to `componentsGroupMapping`
                               
 1.0.0
 ------------------------